Renamed TimeHistHDF to HDFHistoryIO.
